+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix icon alignment in Pagination component
+- Fix Sticky component overflow scrolling position on resize
 
 ## 3.17.0 (September 28, 2023)